+++ CHANGELOG.md
@@ -1,5 +1,4 @@
-# NEXT UP
-
+# 3.4.6
 * Improvements to item matching for Chris's Premades.
 * Zephyr Strike effect improvements.